sense/requirements.txt
# sudo apt-get install sense-hat
paho-mqtt==1.6.1
adafruit-io==2.6.0
protobuf==3.19.4
